remove CelestialAssetRequirement

+++ src/main/java/com/gtnewhorizons/galaxia/outpost/module/ModuleBigHammer.java
@@ -17,9 +17,9 @@ public class ModuleBigHammer extends AutomatedOutpostModule implements IHammer {
     public static final OutpostModuleKind KIND = OutpostModuleKind.BIG_HAMMER;
 
     // spotless:off
-    public final static Map<ItemStack, Integer> constructionCost = new HashMap<ItemStack, Integer>() {{
-        put(new ItemStack(Items.diamond), 8);
-        put(new ItemStack(Items.gold_ingot), 64);
+    public final static Map<ItemStack, Long> constructionCost = new HashMap<ItemStack, Long>() {{
+        put(new ItemStack(Items.diamond), 8L);
+        put(new ItemStack(Items.gold_ingot), 64L);
     }};
     // spotless:on
 
@@ -48,7 +48,7 @@ public class ModuleBigHammer extends AutomatedOutpostModule implements IHammer {
     }
 
     @Override
-    public Map<ItemStack, Integer> getConstructionCost() {
+    public Map<ItemStack, Long> getConstructionCost() {
         return constructionCost;
     }
